[TASK] Add hint to remove SubRoute definition after deactivating package
When a SubRoute definition refers to a non-existing or disabled
package an exception is thrown, which confused some followers of
the mini tutorial of the Welcome screen.

This change adds a notice to remove the SubRoute definition
after deactivating the Welcome package.

Besides this renames "Remove the Welcome package ..." to
"Disable the Welcome package ...".

Note: Some infos have been added to the exception wiki page at
http://wiki.typo3.org/wiki/Exception1318414040

Change-Id: Ifc851f3208eaf3a751747ad678a624b5586b92d7
Resolves: #31284
Releases: 1.0, 1.1